refactor(solver): relocate declaration_projection under operations/ (root-file policy)
The Sound Mode declaration-boundary projection module was added at the
solver crate root, but tsz-solver already exceeds the 4 non-facade
root-file budget, so a new root module trips the root-file policy.

Relocate it into operations/ (it is a polarity-aware type transform: the
type-operations brain that takes TypeId and returns projected TypeIds),
and re-export it from lib.rs as crate::declaration_projection so the
checker keeps the tsz_solver::declaration_projection path. Zero new
non-facade root files.
