+++ crates/tsz-solver/src/operations/declaration_projection.rs
